# Mo-Sys/MoSysVP_5.6.0/Plugins/MoSysVPPro/Content/OnSetVFX/Remote/swagger_server/models/level.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Level(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, level: str=None):  # noqa: E501
        """Level - a model defined in Swagger

        :param level: The level of this Level.  # noqa: E501
        :type level: str
        """
        self.swagger_types = {
            'level': str
        }

        self.attribute_map = {
            'level': 'level'
        }

        self._level = level

    @classmethod
    def from_dict(cls, dikt) -> 'Level':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The level of this Level.  # noqa: E501
        :rtype: Level
        """
        return util.deserialize_model(dikt, cls)

    @property
    def level(self) -> str:
        """Gets the level of this Level.


        :return: The level of this Level.
        :rtype: str
        """
        return self._level

    @level.setter
    def level(self, level: str):
        """Sets the level of this Level.


        :param level: The level of this Level.
        :type level: str
        """

        self._level = level