# Mo-Sys/MoSysVP_5.6.0/Plugins/MoSysVPPro/Content/OnSetVFX/Remote/swagger_server/models/play.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Play(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, play: bool=None):  # noqa: E501
        """Play - a model defined in Swagger

        :param play: The play of this Play.  # noqa: E501
        :type play: bool
        """
        self.swagger_types = {
            'play': bool
        }

        self.attribute_map = {
            'play': 'play'
        }

        self._play = play

    @classmethod
    def from_dict(cls, dikt) -> 'Play':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The play of this Play.  # noqa: E501
        :rtype: Play
        """
        return util.deserialize_model(dikt, cls)

    @property
    def play(self) -> bool:
        """Gets the play of this Play.


        :return: The play of this Play.
        :rtype: bool
        """
        return self._play

    @play.setter
    def play(self, play: bool):
        """Sets the play of this Play.


        :param play: The play of this Play.
        :type play: bool
        """

        self._play = play