# Mo-Sys/MoSysVP_5.6.0/Plugins/MoSysVPPro/Content/OnSetVFX/Remote/swagger_server/models/record.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Record(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, record: bool=None):  # noqa: E501
        """Record - a model defined in Swagger

        :param record: The record of this Record.  # noqa: E501
        :type record: bool
        """
        self.swagger_types = {
            'record': bool
        }

        self.attribute_map = {
            'record': 'record'
        }

        self._record = record

    @classmethod
    def from_dict(cls, dikt) -> 'Record':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The record of this Record.  # noqa: E501
        :rtype: Record
        """
        return util.deserialize_model(dikt, cls)

    @property
    def record(self) -> bool:
        """Gets the record of this Record.


        :return: The record of this Record.
        :rtype: bool
        """
        return self._record

    @record.setter
    def record(self, record: bool):
        """Sets the record of this Record.


        :param record: The record of this Record.
        :type record: bool
        """

        self._record = record