# Mo-Sys/MoSysVP_5.6.0/Plugins/MoSysVPPro/Content/OnSetVFX/Remote/swagger_server/models/restart.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Restart(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, restart: bool=None):  # noqa: E501
        """Restart - a model defined in Swagger

        :param restart: The restart of this Restart.  # noqa: E501
        :type restart: bool
        """
        self.swagger_types = {
            'restart': bool
        }

        self.attribute_map = {
            'restart': 'restart'
        }

        self._restart = restart

    @classmethod
    def from_dict(cls, dikt) -> 'Restart':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The restart of this Restart.  # noqa: E501
        :rtype: Restart
        """
        return util.deserialize_model(dikt, cls)

    @property
    def restart(self) -> bool:
        """Gets the restart of this Restart.


        :return: The restart of this Restart.
        :rtype: bool
        """
        return self._restart

    @restart.setter
    def restart(self, restart: bool):
        """Sets the restart of this Restart.


        :param restart: The restart of this Restart.
        :type restart: bool
        """

        self._restart = restart