# Mo-Sys/MoSysVP_5.6.0/Plugins/MoSysVPPro/Content/OnSetVFX/Remote/swagger_server/models/status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Status(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, status: bool=None):  # noqa: E501
        """Status - a model defined in Swagger

        :param status: The status of this Status.  # noqa: E501
        :type status: bool
        """
        self.swagger_types = {
            'status': bool
        }

        self.attribute_map = {
            'status': 'status'
        }

        self._status = status

    @classmethod
    def from_dict(cls, dikt) -> 'Status':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The status of this Status.  # noqa: E501
        :rtype: Status
        """
        return util.deserialize_model(dikt, cls)

    @property
    def status(self) -> bool:
        """Gets the status of this Status.


        :return: The status of this Status.
        :rtype: bool
        """
        return self._status

    @status.setter
    def status(self, status: bool):
        """Sets the status of this Status.


        :param status: The status of this Status.
        :type status: bool
        """

        self._status = status